anagram_021.py
Two words are anagrams if the letters of one word can be rearranged to form the other word. For example angel and glean are anagrams. Write a Python program called anagram_021.py that reads in pairs of words (one pair per line) from stdin and prints True if the pair are anagrams and False otherwise.